airflow/requirements.txt

apache-airflow==2.7.3
kafka-python==2.0.2
vnstock
pandas
